Switch to SHA-256 + RSA-2048 by default & provide tests
Providing new parameters --keylength and --hash that allows to provide
the length of private keys and the strength of the digest algorithm used
for signing certificates.

Previous hardcoded setting was SHA-1 for the hash, and there was no option
for the key length (although it seems that openssl defaulted to RSA-1024).

Also providing a battery of tests that can be run through
./tests.sh

Resolves: LE-9792
Change-Id: Id0976d4ade5e2ad3bfda04f4de6e3d81d483e0a9