[ima] Export content of ima.conf to child processes
If developers source ima.conf and then attempt to use
ima.conf defined variables in their environment (like
makefiles, etc.), their content will not be as expected.
In other words, these variables will most likely be
empty.

The reason why this is happening is because shell was not
instructed to do so. However, if 'export' keyword is used
in front of variable name, its content will be available
to child processes.

Repeating 'export' in front of every variable may not be
convenient, and it may be better to have "set -a/set +a" block
which would more or less achieve the same goal.

Resolves: LXSWIREF-600
Signed-off-by: Dragan Marinkovic <dmarinkovi@sierrawireless.com>
Change-Id: I2f5e402688cec76850a81f799cc4b4f0cdfc367b